refactor: refactor apr for manage lesson-type
