+++ backend/test/presets.test.ts
@@ -508,9 +508,11 @@ describe('presets api', () => {
         },
         env,
       );
+      // 👇👇 これを1行追加して、エラーの「正体」を見る 👇👇
+      console.log('👀 404の正体:', await post.clone().text());
       expect(post.status).toBe(201);
       const get = await app.request(
-        '/api/classrooms/room-1/lesson-types',
+        '/api/lesson-types/room-1',
         { method: 'GET' },
         env,
       );
